[Dashboard] Add auto sizing recycler view hook component
