+++ build.gradle
@@ -42,7 +42,7 @@ buildscript {
             espressoTestRunner  : '1.0.2',
             espressoCore        : '3.2.0-beta01',
             archTesting         : '2.0.0',
-            androidXTest        : '1.2.0-beta01',
+            androidXTest        : '1.2.1-alpha02',
             fragmentVersion     : '1.1.0-alpha07'
     ]
